ml/requirements.txt
torch
sahi
unidecode
fastapi
uvicorn[standard]
opencv-python
ultralytics
shapely>=1.8.0
tqdm>=4.48.2
pillow>=8.2.0
pybboxes==0.1.6
pyyaml
fire
terminaltables
requests
click